refactor(llm): move parsing options to common protocols
